requirement.txt
pytest
pytest-ordering
pytest-xdist
pytest-rerunfailures
pytest-html
allure-pytest
PyYAML
requests